Cleaned up code, removed Boost dependency (all C++11 now), started experimenting with 3D rendering in SFML and libnoise for procedural texture generation.

+++ Server/src/utility.cpp
@@ -1,9 +1,9 @@
 #include <sstream>
-#include <iostream>
+#include <config.hpp>
 
 #include <SFML/System.hpp>
 
-#include "utility.h"
+#include <utility.hpp>
 
 sf::Clock sfclock;
 
@@ -18,7 +18,7 @@ const std::string GetTimeString() {
   return "[" + ss.str() + "]";
 }
 
-float GetTime() {
+unsigned int GetTime() {
   return sfclock.GetElapsedTime();
 }
